requirements.txt
Flask==2.2.2
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-exporter-otlp==1.19.0

